backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
Werkzeug==3.0.1

# ML/AI Libraries
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4
joblib==1.3.2

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78
mediapipe==0.10.9

# Data validation
marshmallow==3.20.1

# Password hashing
bcrypt==4.1.2
